+++ zerver/lib/cache.py
@@ -268,10 +268,16 @@ def cache_save_user_profile(user_profile):
     # type: (Any) -> None
     cache_set(user_profile_by_id_cache_key(user_profile.id), user_profile, timeout=3600*24*7)
 
+active_user_dict_fields = ['id', 'full_name', 'short_name', 'email', 'is_realm_admin', 'is_bot']
 def active_user_dicts_in_realm_cache_key(realm):
     # type: (Any) -> str
     return "active_user_dicts_in_realm:%s" % (realm.id,)
 
+active_bot_dict_fields = ['id', 'full_name', 'short_name',
+                          'email', 'default_sending_stream__name',
+                          'default_events_register_stream__name',
+                          'default_all_public_streams', 'api_key',
+                          'bot_owner__email', 'avatar_source']
 def active_bot_dicts_in_realm_cache_key(realm):
     # type: (Any) -> str
     return "active_bot_dicts_in_realm:%s" % (realm.id,)
@@ -302,16 +308,16 @@ def flush_user_profile(sender, **kwargs):
     update_user_profile_caches([user_profile])
 
     # Invalidate our active_users_in_realm info dict if any user has changed
-    # name or email
+    # the fields in the dict or become (in)active
     if kwargs.get('update_fields') is None or \
-        len(set(['full_name', 'short_name', 'email', 'is_active']) & set(kwargs['update_fields'])) > 0:
+        len(set(active_user_dict_fields + ['is_active']) & set(kwargs['update_fields'])) > 0:
         cache_delete(active_user_dicts_in_realm_cache_key(user_profile.realm))
 
-    # Invalidate our active_bots_in_realm info dict if any bot has changed
-    bot_fields = {'full_name', 'api_key', 'avatar_source',
-                  'default_all_public_streams', 'is_active',
-                  'default_sending_stream', 'default_events_register_stream'}
-    if user_profile.is_bot and (kwargs['update_fields'] is None or bot_fields & set(kwargs['update_fields'])):
+    # Invalidate our active_bots_in_realm info dict if any bot has
+    # changed the fields in the dict or become (in)active
+    if user_profile.is_bot and (kwargs['update_fields'] is None or
+                                (set(active_bot_dict_fields + ['is_active']) &
+                                 set(kwargs['update_fields']))):
         cache_delete(active_bot_dicts_in_realm_cache_key(user_profile.realm))
 
     # Invalidate realm-wide alert words cache if any user in the realm has changed