+++ zerver/lib/test_helpers.py
@@ -327,8 +327,8 @@ class AuthedTestCase(TestCase):
         return self.assertTrue(actual_count <= count,
                                "len(%s) == %s, > %s" % (queries, actual_count, count))
 
-    def assert_json_error_contains(self, result, msg_substring):
-        self.assertIn(msg_substring, self.get_json_error(result))
+    def assert_json_error_contains(self, result, msg_substring, status_code=400):
+        self.assertIn(msg_substring, self.get_json_error(result, status_code=status_code))
 
     def fixture_data(self, type, action, file_type='json'):
         return open(os.path.join(os.path.dirname(__file__),