+++ zerver/management/commands/logout_all_users.py
@@ -5,7 +5,7 @@ from optparse import make_option
 from django.core.management.base import BaseCommand
 
 from zerver.lib.actions import delete_all_user_sessions, \
-    delete_realm_user_sessions
+    delete_realm_user_sessions, delete_all_deactivated_user_sessions
 from zerver.models import get_realm
 
 class Command(BaseCommand):
@@ -17,11 +17,17 @@ class Command(BaseCommand):
                     action='store',
                     default=None,
                     help="Only logout all users in a particular realm"),
+        make_option('--deactivated-only',
+                    action='store_true',
+                    default=False,
+                    help="Only logout all users who are deactivated"),
         )
 
     def handle(self, *args, **options):
         if options["realm"]:
             realm = get_realm(options["realm"])
             delete_realm_user_sessions(realm)
+        elif options["deactivated_only"]:
+            delete_all_deactivated_user_sessions()
         else:
             delete_all_user_sessions()