+++ docs/changelog.md
@@ -8,6 +8,13 @@ All notable changes to the Zulip server are documented in this file.
 - Added caching of virtualenvs in development.
 - Fixed missing helper scripts for RabbitMQ Nagios plugins.
 
+### 1.3.12 - 2016-05-10
+- CVE-2016-4426: Bot API keys were accessible to other users in the same realm.
+- CVE-2016-4427: Deactivated users could access messages if SSO was enabled.
+- Fixed a RabbitMQ configuration bug that resulted in reordered messages.
+- Added expansive test suite for authentication backends and decorators.
+- Added an option to logout_all_users to delete only sessions for deactivated users.
+
 ### 1.3.11 - 2016-05-02
 - Moved email digest support into the default Zulip production configuration.
 - Added options for configuring Postgres, RabbitMQ, Redis, and memcached